feat: edit plants on dedicated page
